Add 1x1 room prefabs
Part 1 of the maze generation overhaul. It's way more efficient
to instantiate the maze cells using prefabs rather than building
each cell tile by tile. Each prefab has a painstakingly arranged
hierarchy. This will allow the script to pick and choose which
part of the prefab it wants to modify, making the code a lot
easier on the eyes.